Revised string interpolation docs
