+++ docs/csharp/language-reference/tokens/snippets/string-interpolation.cs
@@ -10,18 +10,25 @@ public class StringInterpolation
         Console.WriteLine();
         DelimitInterpolatedExpression();
         Console.WriteLine();
+        InterpolatedRawStringLiteral();
+        Console.WriteLine();
+        InterpolatedRawStringLiteralWithBraces();
+        Console.WriteLine();
         CreateCultureSpecificResults();
+        Console.WriteLine();
+        CreateCultureSpecificResultsOld();
         Console.WriteLine();
         MultiLineExpression();
         Console.WriteLine();
+
         RawStrings.Examples();
         Console.WriteLine();
     }
 
     private static void CompareWithCompositeFormatting()
     {
-        // <Snippet1>
-        string name = "Mark";
+        // <CompareWithCompositeFormatting>
+        var name = "Mark";
         var date = DateTime.Now;
 
         // Composite formatting:
@@ -30,66 +37,90 @@ public class StringInterpolation
         Console.WriteLine($"Hello, {name}! Today is {date.DayOfWeek}, it's {date:HH:mm} now.");
         // Both calls produce the same output that is similar to:
         // Hello, Mark! Today is Wednesday, it's 19:40 now.
-        // </Snippet1>
+        // </CompareWithCompositeFormatting>
     }
 
     private static void AlignAndSpecifyFormat()
     {
-        // <Snippet2>
+        // <AlignAndSpecifyFormat>
         Console.WriteLine($"|{"Left",-7}|{"Right",7}|");
 
         const int FieldWidthRightAligned = 20;
         Console.WriteLine($"{Math.PI,FieldWidthRightAligned} - default formatting of the pi number");
         Console.WriteLine($"{Math.PI,FieldWidthRightAligned:F3} - display only three decimal digits of the pi number");
-        // Expected output is:
+        // Output is:
         // |Left   |  Right|
         //     3.14159265358979 - default formatting of the pi number
         //                3.142 - display only three decimal digits of the pi number
-        // </Snippet2>
+        // </AlignAndSpecifyFormat>
     }
 
     private static void DelimitInterpolatedExpression()
     {
-        // <Snippet3>
+        // <BraceAndConditional>
         string name = "Horace";
         int age = 34;
         Console.WriteLine($"He asked, \"Is your name {name}?\", but didn't wait for a reply :-{{");
         Console.WriteLine($"{name} is {age} year{(age == 1 ? "" : "s")} old.");
-        // Expected output is:
+        // Output is:
         // He asked, "Is your name Horace?", but didn't wait for a reply :-{
         // Horace is 34 years old.
-        // </Snippet3>
+        // </BraceAndConditional>
     }
 
     private static void CreateCultureSpecificResults()
     {
-        // <Snippet4>
+        var currentCulture = System.Globalization.CultureInfo.CurrentCulture;
+        
+        // <CultureSpecific>
         double speedOfLight = 299792.458;
-        FormattableString message = $"The speed of light is {speedOfLight:N3} km/s.";
 
         System.Globalization.CultureInfo.CurrentCulture = System.Globalization.CultureInfo.GetCultureInfo("nl-NL");
-        string messageInCurrentCulture = message.ToString();
+        string messageInCurrentCulture = $"The speed of light is {speedOfLight:N3} km/s.";
 
         var specificCulture = System.Globalization.CultureInfo.GetCultureInfo("en-IN");
-        string messageInSpecificCulture = message.ToString(specificCulture);
+        string messageInSpecificCulture = string.Create(
+            specificCulture, $"The speed of light is {speedOfLight:N3} km/s.");
 
-        string messageInInvariantCulture = FormattableString.Invariant(message);
+        string messageInInvariantCulture = string.Create(
+            System.Globalization.CultureInfo.InvariantCulture, $"The speed of light is {speedOfLight:N3} km/s.");
 
         Console.WriteLine($"{System.Globalization.CultureInfo.CurrentCulture,-10} {messageInCurrentCulture}");
         Console.WriteLine($"{specificCulture,-10} {messageInSpecificCulture}");
         Console.WriteLine($"{"Invariant",-10} {messageInInvariantCulture}");
-        // Expected output is:
+        // Output is:
         // nl-NL      The speed of light is 299.792,458 km/s.
         // en-IN      The speed of light is 2,99,792.458 km/s.
         // Invariant  The speed of light is 299,792.458 km/s.
-        // </Snippet4>
+        // </CultureSpecific>
+        
+        System.Globalization.CultureInfo.CurrentCulture = currentCulture;
+    }
+
+    private static void CreateCultureSpecificResultsOld()
+    {
+        // <CultureSpecificByFormattableString>
+        double speedOfLight = 299792.458;
+        FormattableString message = $"The speed of light is {speedOfLight:N3} km/s.";
+
+        var specificCulture = System.Globalization.CultureInfo.GetCultureInfo("en-IN");
+        string messageInSpecificCulture = message.ToString(specificCulture);
+        Console.WriteLine(messageInSpecificCulture);
+        // Output:
+        // The speed of light is 2,99,792.458 km/s.
+
+        string messageInInvariantCulture = FormattableString.Invariant(message);
+        Console.WriteLine(messageInInvariantCulture);
+        // Output is:
+        // The speed of light is 299,792.458 km/s.
+        // </CultureSpecificByFormattableString>
     }
 
     private static void MultiLineExpression()
     {
         int safetyScore = 42;
 
-        // <Newlines>
+        // <MultiLineExpression>
         string message = $"The usage policy for {safetyScore} is {
             safetyScore switch
             {
@@ -100,32 +131,34 @@ public class StringInterpolation
                 _ => "Issues must be addressed before continued use",
             }
             }";
-        // </Newlines>
+        // </MultiLineExpression>
         Console.WriteLine(message);
     }
 
-    private static void InterpolatedRawLiteralStrings()
+    private static void InterpolatedRawStringLiteral()
     {
-        // <RawInterpolatedLiteralString>
+        // <InterpolatedRawStringLiteral>
         int X = 2;
         int Y = 3;
 
-        var pointMessage = $"""The point "{X}, {Y}" is {Math.Sqrt(X * X + Y * Y)} from the origin""";
+        var pointMessage = $"""The point "{X}, {Y}" is {Math.Sqrt(X * X + Y * Y):F3} from the origin""";
 
         Console.WriteLine(pointMessage);
-        // output:  The point "2, 3" is 3.605551275463989 from the origin.
-        // </RawInterpolatedLiteralString>
-
+        // Output is:
+        // The point "2, 3" is 3.606 from the origin
+        // </InterpolatedRawStringLiteral>
     }
-    private static void InterpolatedRawLiteralStringsWithBraces()
+    
+    private static void InterpolatedRawStringLiteralWithBraces()
     {
-        // <RawInterpolatedLiteralStringWithBraces>
+        // <InterpolatedRawStringLiteralWithBraces>
         int X = 2;
         int Y = 3;
 
-        var pointMessage = $$"""The point {{{X}}, {{Y}}} is {{Math.Sqrt(X * X + Y * Y)}} from the origin""";
+        var pointMessage = $$"""{The point {{{X}}, {{Y}}} is {{Math.Sqrt(X * X + Y * Y):F3}} from the origin}""";
         Console.WriteLine(pointMessage);
-        // output:  The point {2, 3} is 3.605551275463989 from the origin.
-        // </RawInterpolatedLiteralStringWithBraces>
+        // Output is:
+        // {The point {2, 3} is 3.606 from the origin}
+        // </InterpolatedRawStringLiteralWithBraces>
     }
 }
